# Vectorize ProbabilisticMap.IndexOfAny