… src/libraries/System.Private.CoreLib/src/System/Runtime/Intrinsics/Vector256.cs
… src/libraries/System.Private.CoreLib/src/System/SpanHelpers.Char.cs
… src/libraries/System.Private.CoreLib/src/System/String.Manipulation.cs