#### AVX2

|             Method | Toolchain | NewLineFrequency | Length |       Mean |     Error | Ratio |
|------------------- |---------- |----------------- |------- |-----------:|----------:|------:|
| ReplaceLineEndings |      main |                0 |  10000 |  11.481 us | 0.0569 us |  1.00 |
| ReplaceLineEndings |        pr |                0 |  10000 |   1.086 us | 0.0026 us |  0.09 |
|                    |           |                  |        |            |           |       |
| ReplaceLineEndings |      main |             0.05 |  10000 |  24.280 us | 0.0478 us |  1.00 |
| ReplaceLineEndings |        pr |             0.05 |  10000 |  13.983 us | 0.0596 us |  0.58 |
|                    |           |                  |        |            |           |       |
| ReplaceLineEndings |      main |              0.1 |  10000 |  35.616 us | 0.1959 us |  1.00 |
| ReplaceLineEndings |        pr |              0.1 |  10000 |  27.322 us | 0.3056 us |  0.77 |
|                    |           |                  |        |            |           |       |
| ReplaceLineEndings |      main |              0.2 |  10000 |  58.372 us | 1.0346 us |  1.00 |
| ReplaceLineEndings |        pr |              0.2 |  10000 |  52.060 us | 0.2331 us |  0.89 |
|                    |           |                  |        |            |           |       |
| ReplaceLineEndings |      main |                1 |  10000 | 160.760 us | 1.1764 us |  1.00 |
| ReplaceLineEndings |        pr |                1 |  10000 | 201.541 us | 4.8189 us |  1.26 |

#### ARM64

| Method | Length |      Mean |     Error |
|------- |------- |----------:|----------:|
|    Old |  10000 | 13.122 µs | 0.0117 µs |
|    New |  10000 |  3.701 µs | 0.0012 µs |

<details>
<summary>Benchmark source</summary>

```c#
public class ReplaceLineEndingsBenchmark
{
    private string _input;

    [Params(0.0, 0.05, 0.1, 0.2, 1.0)]
    public double NewLineFrequency;

    [Params(10_000)]
    public int Length;

    [GlobalSetup]
    public void Setup()
    {
        char[] input = new char[Length];

        var rng = new Random(42);

        for (int i = 0; i < input.Length; i++)
        {
            if (rng.NextDouble() < NewLineFrequency)
            {
                input[i] = '\n';
            }
            else
            {
                char c;
                do
                {
                    c = (char)rng.Next(0, 65536);
                }
                while ("\r\n\f\u0085\u2028\u2029".Contains(c));

                input[i] = c;
            }
        }

        _input = new string(input);
    }

    [Benchmark]
    public string ReplaceLineEndings() => _input.ReplaceLineEndings();
}
```

</details>

If we care, we could also do this for `LastIndexOfAny`.

#80297 could be interesting for ARM if we could do a 256-bit lookup instead of blending together smaller lookups.