@MihaZupan 2023-01-26
Added `Vector{128/256}.LoadUnsafe(ref char)` and `Vector128.ShuffleUnsafe` as internal APIs for now and replaced a few existing uses. I intentionally excluded the shuffle helper in [`Base64Decoder`](https://github.com/dotnet/runtime/blob/8ba2ad4f46e8011331d29dacf416a8647caa0bbb/src/libraries/System.Private.CoreLib/src/System/Buffers/Text/Base64Decoder.cs#L482-L494) as it feels more readable as-is.