@lewing 2023-02-08
@radekdoulik we should add PackedSimd versions of the Unsafe* functions wasm supports to ease handling all the calling callers.  I'm not sure we have a great pattern to simplify the *.IsHardwareAccelerated paths that works for all the cases yet but that is worth considering too.